% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_apply_amount.R
\name{fert_apply_amount}
\alias{fert_apply_amount}
\title{How much fertilizer do I need?}
\usage{
fert_apply_amount(sqft, nutrient_percent, desired_pounds)
}
\arguments{
\item{sqft}{numeric - the total square feet of your lawn.}

\item{nutrient_percent}{numeric - the percentage of the nutrient to apply.}

\item{desired_pounds}{numeric  - desired pounds of nutrient per one thousand square feet}
}
\value{
The total pounds of fertilizer, in pounds, to apply to your lawn.
}
\description{
Given the nutrient percentage in a bag of fertilizer,
calculate the total pounds of fertilizer required to achieve a specific
nutrient target per one thousand square feet.
}
\examples{
fert_apply_amount(5000, 24, 1) # 21 pounds
}
