% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_bag_rate.R
\name{fert_bag_rate}
\alias{fert_bag_rate}
\title{Analyze the nutrient content of a bag of fertilizer}
\usage{
fert_bag_rate(sqft, n, p, k, pounds)
}
\arguments{
\item{sqft}{numeric - the total square feet of your lawn.}

\item{n}{numeric - the percentage of the nitrogen.}

\item{p}{numeric - the percentage of the phosphorus.}

\item{k}{numeric - the percentage of the potassium}
}
\value{
A list containing the amount of nitrogen, phosphorus, and potassium
 that will be applied to your lawn.
}
\description{
Calculate the pounds per one thousand square feet of
each nutrient in a bag of fertilizer.
}
\examples{
fert_bag_rate(5000,24, 25, 4, 15)
}
