% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_liquid_rate.R
\name{fert_liquid_rate}
\alias{fert_liquid_rate}
\title{Liquid fertilizer rates per one thousand}
\usage{
fert_liquid_rate(nutrient_percent, weight_in_pounds, ounces)
}
\arguments{
\item{nutrient_percent}{numeric - the percent of nutrient in the liquid fertilizer.}

\item{weight_in_pounds}{numeric - the weight of the liquid in pounds.}

\item{ounces}{numeric - the total of ounces for calculating the nutrient value for ounce.}
}
\value{
The total pounds per thousand for each ounce of liquid fertilizer applied.
}
\description{
Calculate the amount of nutrient, per ounce, in a liquid fertilizer.
}
\examples{
fert_liquid_rate(5, 2.5, 32)
}
