% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_per_1000.R
\name{nutrient_per_1000}
\alias{nutrient_per_1000}
\title{Fertilizer rates per one thousand square feet}
\usage{
nutrient_per_1000(sqft, nutrient_percent, pounds)
}
\arguments{
\item{sqft}{numeric - the total square feet of your lawn.}

\item{nutrient_percent}{numeric - the percentage of the nutrient to apply.}

\item{pounds}{numeric  - the total weight, in pounds, of the fertilizer bag.}
}
\value{
The total pounds of a specific nutrient that will be applied to your lawn.
}
\description{
Given the percentage and pounds of a specific nutrient, calculate the
pounds per one thousand square feet that will be applied
to your lawn.
}
\examples{
nutrient_per_1000(5000, 24, 15) # 0.72 lbs. per 1,000 sqft
}
