% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_rate.R
\name{release_rate}
\alias{release_rate}
\title{Analyze the release rate of fertilizer}
\usage{
release_rate(total_nutrient, slow_release, quick_release)
}
\arguments{
\item{total_nutrient}{numeric - the total ponds per thousand of nutrient being applied.}

\item{slow_release}{numeric - the percentage of the nutrient value that is slow release.}

\item{quick_release}{numeric - the percentage of the nutrient value that is quick release.}
}
\value{
A list containing the amount of nutrient in pounds per thousand
that will be released slowly and quickly.
}
\description{
Calculate the amount of slow and quick release nutrients based on the
total pounds of a specific nutrient being applied.
}
\examples{
release_rate(1, 7, 17) #24\%N 7\% slow release, 17\% quick release
}
